Minimize the use of `BoardLoader`.
