+++ packages/breadboard/src/kits/load.ts
@@ -6,7 +6,7 @@
 
 import { inspect } from "../inspector/index.js";
 import { loadWithFetch } from "../loader/default.js";
-import { BoardLoader } from "../loader/loader.js";
+import { createLoader } from "../loader/index.js";
 import { BoardRunner } from "../runner.js";
 import { GraphDescriptor, Kit, KitManifest, NodeHandler } from "../types.js";
 import { asRuntimeKit } from "./ctors.js";
@@ -19,8 +19,11 @@ const getGraphDescriptor = async (
   entry: ManifestEntry
 ) => {
   if (typeof entry === "string") {
-    const loader = new BoardLoader({});
+    const loader = createLoader();
     const result = await loader.load(new URL(entry, base));
+    if (result === null) {
+      throw new Error(`Unable to load graph descriptor from "${entry}"`);
+    }
     return result;
   } else if (entry.edges && entry.nodes) {
     const url = new URL(base);
