+++ packages/breadboard/src/loader/loader.ts
@@ -6,12 +6,13 @@
 
 import type { GraphDescriptor, SubGraphs } from "../types.js";
 import type { GraphProvider, GraphLoader } from "./types.js";
-import { SENTINEL_BASE_URL, createLoader } from "./index.js";
+import { DefaultGraphProvider } from "./default.js";
+
+export const SENTINEL_BASE_URL = new URL("sentinel:///");
 
 export type BoardLoaderArguments = {
   graphs?: SubGraphs;
   graphProviders?: GraphProvider[];
-  loader?: GraphLoader;
 };
 
 export type BoardLoaderType = "hash" | "other";
@@ -28,36 +29,58 @@ export const sameWithoutHash = (a: URL, b: URL): boolean => {
 
 export class BoardLoader implements GraphLoader {
   #graphs?: SubGraphs;
-  #loader: GraphLoader;
+  #graphProviders: GraphProvider[];
 
-  constructor({ graphs, graphProviders, loader }: BoardLoaderArguments) {
+  constructor({ graphs, graphProviders }: BoardLoaderArguments) {
     this.#graphs = graphs;
-    this.#loader = loader || createLoader(graphProviders);
+    this.#graphProviders = [
+      ...(graphProviders || []),
+      new DefaultGraphProvider(),
+    ];
   }
 
-  async #loadWithLoader(url: URL): Promise<GraphDescriptor> {
-    const graph = await this.#loader.load(url);
-    if (!graph) {
-      throw new Error(`Unable to load graph from "${url.href}"`);
+  async #loadWithProviders(url: URL): Promise<GraphDescriptor | null> {
+    for (const provider of this.#graphProviders) {
+      const capabilities = provider.canProvide(url);
+      if (capabilities === false) {
+        continue;
+      }
+      if (capabilities.load) {
+        const graph = await provider.load(url);
+        if (graph !== null) {
+          graph.url = url.href;
+          return graph;
+        }
+      }
     }
-    graph.url = url.href;
+    console.warn(`Unable to load graph from "${url.href}"`);
+    return null;
+  }
+
+  async #loadOrWarn(url: URL): Promise<GraphDescriptor | null> {
+    const graph = await this.#loadWithProviders(url);
+    if (!graph) {
+      return null;
+    }
     return graph;
   }
 
-  #getSubgraph(url: URL, subgraphs?: SubGraphs): GraphDescriptor {
+  #getSubgraph(url: URL, subgraphs?: SubGraphs): GraphDescriptor | null {
     const hash = url.hash.substring(1);
     if (!subgraphs) {
-      throw new Error("No sub-graphs to load from");
+      console.warn(`No sub-graphs to load "#${hash}" from`);
+      return null;
     }
     const graph = subgraphs[hash];
     if (!graph) {
-      throw new Error(`No graph found for hash: #${hash}`);
+      console.warn(`No graph found for hash: #${hash}`);
+      return null;
     }
     graph.url = url.href;
     return graph;
   }
 
-  async load(url: URL): Promise<GraphDescriptor> {
+  async load(url: URL): Promise<GraphDescriptor | null> {
     if (url.hash) {
       // This is a bit of a special case: some graphs are constructed in
       // memory, so they don't have a URL to base on. In such cases, the
@@ -65,10 +88,13 @@ export class BoardLoader implements GraphLoader {
       if (sameWithoutHash(url, SENTINEL_BASE_URL)) {
         return this.#getSubgraph(url, this.#graphs);
       } else {
-        const superGraph = await this.#loadWithLoader(removeHash(url));
+        const superGraph = await this.#loadOrWarn(removeHash(url));
+        if (!superGraph) {
+          return null;
+        }
         return this.#getSubgraph(url, superGraph.graphs);
       }
     }
-    return await this.#loadWithLoader(url);
+    return await this.#loadOrWarn(url);
   }
 }
