+++ packages/breadboard/tests/inspector/traversal.ts
@@ -7,14 +7,14 @@
 // These test traversal of the inspector API.
 
 import test from "ava";
-import { BoardLoader } from "../../src/loader/loader.js";
 import { inspectableGraph } from "../../src/inspector/graph.js";
+import { createLoader } from "../../src/loader/index.js";
 
 const BASE_URL = new URL("../../../tests/inspector/data/", import.meta.url);
 
 const load = async (url: string) => {
   const base = BASE_URL;
-  const loader = new BoardLoader({});
+  const loader = createLoader();
   const result = await loader.load(new URL(url, base));
   if (!result) return undefined;
   return inspectableGraph(result);
